+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+# [v1.3.0](unreleased)
+### Notes
+- This release supports for API 800,1000,1200  for resource listed below.
+
+### Oneview Features supported
+- Hypervisor Manager
+
 # [v1.2.0]
 ### Notes
 - This release supports API1200 minimally where we can use OneView API1200 with this SDK. No new fields are added/deleted to support API1200. Complete support will be done in next releases.